# docs/api.yml
openapi: 3.1.0
...
info:
  title: yomitai-api-doc
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
...
  '/todo/{todoId}':
    parameters:
      - schema:
          type: string
        name: todoId
        in: path
        required: true
    get:
      summary: GET TodoById
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  todo:
                    $ref: '#/components/schemas/Todo'
                    x-stoplight:
                      id: bc1sjsqfobpwb
      operationId: get-todo-todoId
      description: |
        idを元に、Todo詳細を取得
      parameters:
        - $ref: '#/components/parameters/next-auth.session-token'
    put:
      summary: Put TodoById
      operationId: put-todo-todoId
      responses:
        '200':
          description: OK
      description: todoを更新
      parameters:
        - $ref: '#/components/parameters/next-auth.session-token'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  x-stoplight:
                    id: dxwa2riya7fi9
                content:
                  type: string
                  x-stoplight:
                    id: mb05j5nux4khd
                status:
                  $ref: '#/components/schemas/TodoStatus'
                  x-stoplight:
                    id: z7ier7lu5i2mc
              required:
                - id
                - status
    delete:
      summary: Delete TodoById
      operationId: delete-todo-todoId
      responses:
        '200':
          description: OK
      description: todoを削除（アーカイブ）
      parameters:
        - $ref: '#/components/parameters/next-auth.session-token'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  x-stoplight:
                    id: iqu1nqj1wuf61
...
components:
  schemas:
...
    TodoStatus:
      title: TodoStatus
      x-stoplight:
        id: a25sgm5bbd5k6
      type: string
      enum:
        - TODO
        - COMPLETED
...
    Todo:
      title: Todo
      x-stoplight:
        id: iv99lfg3s2hpa
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          x-stoplight:
            id: cif6ejvk3mgnl
        content:
          type: string
          x-stoplight:
            id: 7qi6v94b0dmht
        imageUrl:
          type: string
          x-stoplight:
            id: yngt7jwmkxpn2
  parameters:
    next-auth.session-token:
      name: next-auth.session-token
      in: cookie
      required: true
      schema:
        type: string
      description: next-authで自動追加されるcokkie情報